Made it so that you can get the PAL_QUIT event.
Both the fake window and the real window used to use
the same window procedure, this caused problems where
if you handled the WM_CLOSE, WM_DESTROY or WM_QUIT message
on one window, the application would close. And the original
window proc did not even handle WM_CLOSE or WM_DESTROY, so we never
get WM_QUIT. This commit fixes that.

Signed-off-by: Abdul Sarhan <abdulrahmanayman1@gmail.com>