Fixed: pal_init() on Linux & Bug Where Linux Dummy Window Was Mapped.
We needed to create a dummy window for some reason.
The window was mapped, which caused it to appear on the screen,
and recieve expose events, not good. And we also flushed it which
is not necessary for a window that is not supposed to be mapped.
And we also needed to init the dummy window before the raw input
because raw input depends on the dummy window, I fixed that too.